Green check for Azure SQL Database as well
As it is says in the contents, this document is partially applies to Azure SQL Database as well. 
"SQL Database enforces password complexity. The password expiration and policy enforcement sections do not apply to SQL Database."

Therefore, we should have the green check for Azure SQL Database at the APPLIES TO section.